refactor(SLB-278): split e2e tests into drupal/decap
So they can be run independently, depending on which CMS is active.
