+++ tests/e2e/package.json
@@ -8,9 +8,12 @@
   "type": "module",
   "scripts": {
     "test:static": "tsc --noEmit && eslint '**/*.{ts,tsx,js,jsx}' --ignore-path='./.gitignore'",
-    "test:integration": "playwright test",
-    "test:headed": "playwright test --headed",
-    "dev": "playwright test --ui",
+    "test:integration:drupal": "playwright test --config playwright.config.drupal.ts",
+    "test:integration:decap": "playwright test --config playwright.config.decap.ts",
+    "test:headed:drupal": "playwright test --headed --config playwright.config.drupal.ts",
+    "test:headed:decap": "playwright test --headed --config playwright.config.decap.ts",
+    "dev:decap": "playwright test --ui --config playwright.config.decap.ts",
+    "dev:drupal": "playwright test --ui --config playwright.config.drupal.ts",
     "webform-snapshots": "playwright test --config playwright.config.webforms.ts",
     "webform-snapshots:headed": "playwright test --config playwright.config.webforms.ts --headed",
     "download-browsers": "playwright install chromium"
