+++ tests/e2e/specs/decap/decap-pages.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@playwright/test';
 
-import { websiteUrl } from '../helpers/url';
+import { websiteUrl } from '../../helpers/url';
 
 test.describe('decap pages', () => {
   test('example decap page is rendered', async ({ page }) => {
