+++ tests/e2e/specs/drupal/campaign-urls.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@playwright/test';
 
-import { websiteUrl } from '../helpers/url';
+import { websiteUrl } from '../../helpers/url';
 
 test.describe('campaign url redirects', () => {
   test('simple redirects', async ({ page }) => {
