+++ tests/e2e/specs/drupal/homepage.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@playwright/test';
 
-import { websiteUrl } from '../helpers/url';
+import { websiteUrl } from '../../helpers/url';
 
 test.describe('the homepage', () => {
   test('exists in english', async ({ page }) => {
