+++ tests/e2e/specs/drupal/menus.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@playwright/test';
 
-import { websiteUrl } from '../helpers/url';
+import { websiteUrl } from '../../helpers/url';
 
 test.describe('menus', () => {
   test('main navigation', async ({ page }) => {
