+++ tests/e2e/specs/drupal/redirects.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@playwright/test';
 
-import { websiteUrl } from '../helpers/url';
+import { websiteUrl } from '../../helpers/url';
 
 test.describe('drupal redirects', () => {
   test('are detected', async ({ page }) => {
