+++ tests/e2e/specs/drupal/webforms.spec.ts
@@ -1,6 +1,6 @@
 import { expect, test } from '@playwright/test';
 
-import { websiteUrl } from '../helpers/url';
+import { websiteUrl } from '../../helpers/url';
 
 test('Webforms work', async ({ page }) => {
   await page.goto(websiteUrl('/en/blocks-complete'));
